+++ upcloud/request/managed_database_test.go
@@ -429,7 +429,7 @@ func TestCreateManagedDatabaseUserRequest(t *testing.T) {
 		Password:       "new-password",
 		Authentication: upcloud.ManagedDatabaseUserAuthenticationCachingSHA2Password,
 		PGAccessControl: &upcloud.ManagedDatabaseUserPGAccessControl{
-			AllowReplication: true,
+			AllowReplication: upcloud.BoolPtr(true),
 		},
 		RedisAccessControl: &upcloud.ManagedDatabaseUserRedisAccessControl{
 			Categories: &[]string{"+@set"},
@@ -475,7 +475,7 @@ func TestModifyManagedDatabaseUserAccessControlRequest(t *testing.T) {
 		ServiceUUID: "fakeuuid",
 		Username:    "fakeuser",
 		PGAccessControl: &upcloud.ManagedDatabaseUserPGAccessControl{
-			AllowReplication: true,
+			AllowReplication: upcloud.BoolPtr(true),
 		},
 		RedisAccessControl: &upcloud.ManagedDatabaseUserRedisAccessControl{
 			Categories: &[]string{"+@set"},